change so only staff can add/modify/delete models, instruments, or cal events

+++ backend/tables/views.py
@@ -23,6 +23,8 @@ def calibration_event_list(request):
         return get_page_response(calibration_events, request, CalibrationEventReadSerializer, "calibration_events", nextPage, previousPage)
 
     elif request.method == 'POST':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         # get item model, instrument, and user from request
         try:
             vendor = request.data['vendor']
@@ -67,6 +69,8 @@ def calibration_event_detail(request, pk):
         return Response(serializer.data)
 
     elif request.method == 'PUT':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         # fill in immutable fields and grab new user's pk
         request.data['instrument'] = calibration_event.instrument.pk
         if 'username' in request.data:
@@ -86,6 +90,8 @@ def calibration_event_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         calibration_event.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -104,6 +110,8 @@ def instruments_list(request):
         return get_page_response(instruments, request, ListInstrumentReadSerializer, "instruments", nextPage, previousPage)
 
     elif request.method == 'POST':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         # get model pk from vendor and model number
         try:
             vendor = request.data['vendor']
@@ -137,6 +145,8 @@ def instruments_detail(request, pk):
         return Response(serializer.data)
 
     elif request.method == 'PUT':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         if 'vendor' in request.data or 'model_number' in request.data:
             vendor = request.data['vendor'] if 'vendor' in request.data else instrument.item_model.vendor
             model_number = request.data['model_number'] if 'model_number' in request.data else instrument.item_model.model_number
@@ -155,6 +165,8 @@ def instruments_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         instrument.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -173,6 +185,8 @@ def models_list(request):
         return get_page_response(models, request, ItemModelSerializer, "models", nextPage, previousPage)
 
     elif request.method == 'POST':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         serializer = ItemModelSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -197,6 +211,8 @@ def models_detail(request, pk):
         return Response(serializer.data)
 
     elif request.method == 'PUT':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         if 'vendor' not in request.data: request.data['vendor'] = model.vendor
         if 'model_number' not in request.data: request.data['model_number'] = model.model_number
         serializer = ItemModelSerializer(model, data=request.data, context={'request': request})
@@ -206,6 +222,8 @@ def models_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if not request.user.is_staff:
+            return Response("User does not have permission.", status=status.HTTP_401_UNAUTHORIZED)
         if len(model.instrument_set.all()) > 0:
             return Response("Cannot delete model with instrument instances.", status=status.HTTP_400_BAD_REQUEST)
         else:
